# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "econpaper-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/{session_id}/code-export": {
      "get": {
        "summary": "Export Code",
        "description": "导出代码文件。\n\nParameters\n----------\nsession_id : str\n    会话 ID。\nformat : str\n    代码格式：``py`` / ``do`` / ``R`` / ``m``。默认 ``py``。\n\nReturns\n-------\nPlainTextResponse\n    带 ``Content-Disposition: attachment; filename=\"...\"`` 的代码文本。\n\nRaises\n------\nHTTPException\n    - 404: session 不存在，或 session 无 code_translations\n    - 400: 不支持的 format",
        "operationId": "export_code_sessions__session_id__code_export_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "py",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/doc-export": {
      "get": {
        "summary": "Export Document",
        "description": "导出文档：format=tex|pdf|docx，template=4 模板之一。",
        "operationId": "export_document_sessions__session_id__doc_export_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tex",
              "title": "Format"
            }
          },
          {
            "name": "template",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "cn_journal",
              "title": "Template"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/eda": {
      "post": {
        "summary": "Run Eda",
        "description": "Run an EDA action on the session's dataset.\n\nRequest body: ``{\"action\": \"describe\"|\"corr\"|\"plot\"|\"scatter\"|\"regression\"|\"missing\"}``.\n- ``describe`` → ``{columns, rows}`` per-variable stats table.\n- ``corr``     → ``{variables, matrix}`` Pearson correlation matrix.\n- ``missing``  → ``{columns, rows}`` missing-value report.\n- ``plot`` / ``scatter`` / ``regression`` → placeholder (later tickets).\n\nStage D：返回 ``EdaResponse``（``action`` + ``result``，``result`` 承载\n各 action 的具体 shape；``EdaResponse`` 用 ``extra=\"allow\"`` 保留\ndescribe / corr / missing 的原始字段）。",
        "operationId": "run_eda_sessions__session_id__eda_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EdaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EdaRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "EdaRequest",
        "description": "POST /sessions/{id}/eda 请求体。"
      },
      "EdaResponse": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "result": {
            "title": "Result"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "EdaResponse",
        "description": "POST /sessions/{id}/eda 返回体。\n\nEDA 动作返回的 shape 因 action 而异（describe / corr / missing /\nplaceholder），统一用 ``Any`` 承载具体字段，但保留 ``action`` 标识。\ndescribe / corr / missing 的结构由各 action 的实现决定，前端按\n``action`` 分支渲染。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}